Fix margin in body, set up height, rename ToDoit to Home
