+++ to-do-app/src/ToDoApp.css
@@ -79,7 +79,6 @@
   width: 100%;
   overflow: scroll;
   align-items: center;
-  margin-top: 8px;
 }
 
 .Logo-Container{
