+++ to-do-app/src/components/NewList/style.css
@@ -1,0 +1,12 @@
+.new-list {
+  width: 199px;
+  height: 33px;
+  left: 197px;
+  top: 603px;
+  font-style: italic;
+  font-weight: 100;
+  font-size: 28px;
+  line-height: 33px;
+  letter-spacing: 0.01em;
+  color: #FFFFFF;
+}
